updated user routes

+++ Nodejs_Server/controllers/user.controller.js
@@ -5,10 +5,16 @@ import { User }  from "../model/user.model.js";
 const createUser = async(req, res) => {
     try {
         const data = req.body;
-        // TODO: User class is not used here, so we assume the request already has all the required attributes before putting into db?
-        const userdb = firestore.collection('users');
-        await userdb.doc(String(data.id)).set(data);
-        res.send("user added!");
+        if (data.userName == null || data.primaryLevel == null) {res.send("please fill up all fields!");}
+        else {
+            const newUser = new User(data.userName, data.primaryLevel);
+            const userdb = firestore.collection('users');
+            userdb.add(JSON.parse(JSON.stringify(newUser)))
+            .then(function(docRef) {
+                userdb.doc(docRef.id).set({"id": docRef.id}, { merge: true });
+            });
+            res.send("user added!");
+        }
     } catch (error) {
         res.status(400).send(error.message);
         res.send("error adding user!");
@@ -18,10 +24,10 @@ const createUser = async(req, res) => {
 // Retrieve and return a user
 const getUser = async(req, res) => {
     try {
-        const data = req.query.id;
+        const id = req.query.id;
         const userdb = firestore.collection('users');
         const currUser = await userdb.doc(String(id)).get();
-        res.send(currUser.data());
+        res.send(JSON.stringify(currUser.data()));
     } catch (error) {
         res.status(400).send(error.message);
         res.send("user doesnt exist!");
@@ -33,25 +39,33 @@ const deleteUser = async(req, res) => {
     try {
         const id = req.query.id;
         const userdb = firestore.collection('users');
-        await userdb.doc(String(id)).delete();
-        res.send("user is deleted!");
+        const currUser = await userdb.doc(String(id)).get();
+        if (!currUser.exists) {res.send("no such user!");}
+        else {
+            await userdb.doc(String(id)).delete();
+            res.send("user is deleted!");
+        }
+        
     } catch (error) {
-        return res.status(400).send("invalid user");
+        return res.status(400).send("invalid");
         //res.send("invalid user");
     }
 };
 
-// Update a user identified by the userId in the request
+// Update a user identified by the userId in the request body
 const updateUser = async(req, res) => {
     try {
         const data = req.body;
+        const id = data.id;
         const userdb = firestore.collection('users');
-        const updatedUser = await userdb.doc(String(data.id)).set(data)
-        res.send("user is updated!");
+        const currUser = await userdb.doc(String(id)).get();
+        if (!currUser.exists) {res.send("no such user!");}
+        else { const updatedUser = await userdb.doc(String(data.id)).set(data)
+            res.send("user is updated!");}
     } catch (error) {
         res.status(400).send(error.message);
         res.send("invalid update");
-    }
+    } 
 
 };
